+++ DiseaseNet/unconditional_logistic.py
@@ -75,7 +75,7 @@ def logistic_model(d1:float,d2:float,phenotype_df_exposed:pd.DataFrame,id_col,tr
     
     #check var of covariates, remove these with var()==0
     for var in covariates:
-        phenotype_df_exposed[var] = phenotype_df_exposed[var].astype(np.float32)
+        phenotype_df_exposed[var] = phenotype_df_exposed[var].astype(float)
         if phenotype_df_exposed[var].var() <= 0: #lowest var() allowed
             covariates.remove(var)
     
@@ -89,7 +89,7 @@ def logistic_model(d1:float,d2:float,phenotype_df_exposed:pd.DataFrame,id_col,tr
     #simple method
     if method == 'CN':
         try:
-            model = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=np.int16),
+            model = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=int),
                              phenotype_df_exposed[['d1','constant']+covariates].values)
             result = model.fit(disp=False,method='bfgs')
             beta,se,p,aic = result.params[0], result.bse[0],result.pvalues[0],result.aic
@@ -105,7 +105,7 @@ def logistic_model(d1:float,d2:float,phenotype_df_exposed:pd.DataFrame,id_col,tr
             try:
                 #model
                 model_1_vars = ['d1','constant']+all_diseases_var #only disease variables
-                model = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=np.int16),
+                model = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=int),
                                  phenotype_df_exposed[model_1_vars].values)
                 
                 # Initial alphas to check
@@ -122,7 +122,7 @@ def logistic_model(d1:float,d2:float,phenotype_df_exposed:pd.DataFrame,id_col,tr
                 final_best_alpha, final_disease_vars = find_best_alpha_and_vars(model,alpha_range,alpha_lst,model_1_vars)
                 
                 #fit the final model
-                model_final = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=np.int16),
+                model_final = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=int),
                                        phenotype_df_exposed[final_disease_vars+covariates].values)
                 result_final = model_final.fit(disp=False,method='bfgs')
                 beta,se,p,aic = result_final.params[0], result_final.bse[0],result_final.pvalues[0],result_final.aic
@@ -138,14 +138,14 @@ def logistic_model(d1:float,d2:float,phenotype_df_exposed:pd.DataFrame,id_col,tr
             try:
                 #fit the initial model to get the non-zero disease list
                 model_1_vars = ['d1','constant']+all_diseases_var #only disease variables
-                model = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=np.int16),
+                model = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=int),
                                  phenotype_df_exposed[model_1_vars].values)
                 result = model.fit_regularized(method='l1', alpha=alpha_lst*alpha_single, disp=False)
                 non_zero_indices = np.nonzero(result.params != 0)[0]
                 final_disease_vars = [model_1_vars[i] for i in non_zero_indices]
                 
                 #fit the final model
-                model_final = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=np.int16),
+                model_final = sm.Logit(np.asarray(phenotype_df_exposed['d2'],dtype=int),
                                        phenotype_df_exposed[final_disease_vars+covariates].values)
                 result_final = model_final.fit(disp=False,method='bfgs')
                 beta,se,p,aic = result_final.params[0], result_final.bse[0],result_final.pvalues[0],result_final.aic
@@ -163,7 +163,7 @@ def logistic_model(d1:float,d2:float,phenotype_df_exposed:pd.DataFrame,id_col,tr
         try:
             #generate PC from other diseases variables
             pca = PCA(n_components=pca_number)
-            disease_vars_transformed = pca.fit_transform(np.asarray(phenotype_df_exposed[all_diseases_var],dtype=np.float32))
+            disease_vars_transformed = pca.fit_transform(np.asarray(phenotype_df_exposed[all_diseases_var],dtype=float))
             pca_cols = [f'PCA_{i}' for i in range(disease_vars_transformed.shape[1])]
             disease_vars_transformed = pd.DataFrame(disease_vars_transformed,columns=pca_cols)
             disease_vars_transformed.index = phenotype_df_exposed.index
@@ -172,7 +172,7 @@ def logistic_model(d1:float,d2:float,phenotype_df_exposed:pd.DataFrame,id_col,tr
             variance_explained = sum(pca.explained_variance_ratio_)
             
             #fit model with PCA covariates
-            model_final = sm.Logit(np.asarray(phenotype_df_exposed_PCA['d2'],dtype=np.int16),
+            model_final = sm.Logit(np.asarray(phenotype_df_exposed_PCA['d2'],dtype=int),
                                    phenotype_df_exposed_PCA[['d1','constant']+covariates+pca_cols].values)
             result_final = model_final.fit(disp=False,method='bfgs')
             beta,se,p,aic = result_final.params[0], result_final.bse[0],result_final.pvalues[0],result_final.aic
